+++ packages/default-resolver/package.json
@@ -32,12 +32,12 @@
   "homepage": "https://github.com/pnpm/pnpm/blob/master/packages/default-resolver#readme",
   "dependencies": {
     "@pnpm/error": "workspace:1.2.0",
+    "@pnpm/fetch": "workspace:^2.0.2",
     "@pnpm/git-resolver": "workspace:4.0.15",
     "@pnpm/local-resolver": "workspace:5.0.10",
     "@pnpm/npm-resolver": "workspace:8.1.2",
     "@pnpm/resolver-base": "workspace:7.0.3",
-    "@pnpm/tarball-resolver": "workspace:4.0.2",
-    "fetch-from-npm-registry": "workspace:^4.1.2"
+    "@pnpm/tarball-resolver": "workspace:4.0.2"
   },
   "devDependencies": {
     "@pnpm/default-resolver": "link:",
