--- packages/fetch-from-npm-registry/CHANGELOG.md
@@ -1,37 +1,0 @@
-# fetch-from-npm-registry
-
-## 4.1.2
-
-### Patch Changes
-
-- @pnpm/fetch@2.0.2
-
-## 4.1.1
-
-### Patch Changes
-
-- @pnpm/fetch@2.0.1
-
-## 4.1.0
-
-### Minor Changes
-
-- 2ebb7af33: Print a warning when request fails and a retry will happen. Breaking changes in the programmatic API of `@pnpm/fetch`.
-
-### Patch Changes
-
-- Updated dependencies [2ebb7af33]
-  - @pnpm/fetch@2.0.0
-
-## 4.0.3
-
-### Patch Changes
-
-- 872f81ca1: Don't remove authorization headers when redirecting requests to the same host.
-
-## 4.0.3
-
-### Patch Changes
-
-- @pnpm/fetch@1.0.4
-- @pnpm/npm-registry-agent@2.0.4
