--- packages/fetch-from-npm-registry/package.json
@@ -1,44 +1,0 @@
-{
-  "name": "fetch-from-npm-registry",
-  "version": "4.1.2",
-  "description": "A fetch lib specifically for using with the npm registry",
-  "main": "lib/index.js",
-  "typings": "lib/index.d.ts",
-  "files": [
-    "lib",
-    "!*.map"
-  ],
-  "engines": {
-    "node": ">=10.13"
-  },
-  "scripts": {
-    "lint": "tslint -c ../../tslint.json src/**/*.ts test/**/*.ts",
-    "_test": "cd ../.. && c8 --reporter lcov --reports-dir packages/fetch-from-npm-registry/coverage ts-node packages/fetch-from-npm-registry/test --type-check",
-    "test": "pnpm run compile && pnpm run _test",
-    "prepublishOnly": "pnpm run compile",
-    "compile": "rimraf lib tsconfig.tsbuildinfo && tsc --build"
-  },
-  "repository": "https://github.com/pnpm/pnpm/blob/master/packages/fetch-from-npm-registry",
-  "keywords": [
-    "fetch",
-    "npm"
-  ],
-  "author": "Zoltan Kochan <z@kochan.io> (https://www.kochan.io/)",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/pnpm/pnpm/issues"
-  },
-  "homepage": "https://github.com/pnpm/pnpm/blob/master/packages/fetch-from-npm-registry#readme",
-  "peerDependencies": {
-    "@pnpm/logger": "^3.1.0"
-  },
-  "dependencies": {
-    "@pnpm/fetch": "workspace:2.0.2",
-    "@pnpm/npm-registry-agent": "workspace:2.0.3"
-  },
-  "devDependencies": {
-    "fetch-from-npm-registry": "link:",
-    "nock": "12.0.3"
-  },
-  "funding": "https://opencollective.com/pnpm"
-}
