--- packages/fetch-from-npm-registry/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "extends": "@pnpm/tsconfig",
-  "compilerOptions": {
-    "outDir": "lib",
-    "rootDir": "src"
-  },
-  "include": [
-    "src/**/*.ts",
-    "../../typings/**/*.d.ts"
-  ],
-  "references": [
-    {
-      "path": "../fetch"
-    },
-    {
-      "path": "../npm-registry-agent"
-    }
-  ]
-}
