+++ packages/fetch/package.json
@@ -13,7 +13,8 @@
   },
   "scripts": {
     "lint": "tslint -c ../../tslint.json src/**/*.ts test/**/*.ts",
-    "test": "pnpm run compile",
+    "_test": "cd ../.. && c8 --reporter lcov --reports-dir packages/fetch/coverage ts-node packages/fetch/test --type-check",
+    "test": "pnpm run compile && pnpm run _test",
     "prepublishOnly": "pnpm run compile",
     "compile": "rimraf lib tsconfig.tsbuildinfo && tsc --build"
   },
@@ -33,14 +34,17 @@
   "homepage": "https://github.com/pnpm/pnpm/blob/master/packages/fetch#readme",
   "dependencies": {
     "@pnpm/core-loggers": "workspace:^4.1.2",
+    "@pnpm/npm-registry-agent": "workspace:^2.0.3",
     "@zkochan/retry": "^0.2.0",
     "node-fetch": "2.6.0",
     "node-fetch-unix": "2.3.0"
   },
   "devDependencies": {
+    "@pnpm/fetch": "link:",
     "@pnpm/logger": "^3.2.2",
     "@types/node-fetch": "^2.5.7",
-    "cpy-cli": "^3.1.1"
+    "cpy-cli": "^3.1.1",
+    "nock": "^13.0.0"
   },
   "funding": "https://opencollective.com/pnpm"
 }
