+++ packages/fetch/test/fetchFromRegistry.ts
@@ -1,10 +1,10 @@
 ///<reference path="../../../typings/index.d.ts"/>
-import createRegClient from 'fetch-from-npm-registry'
+import { createFetchFromRegistry } from '@pnpm/fetch'
 import nock = require('nock')
 import test = require('tape')
 
 test('fetchFromNpmRegistry', async t => {
-  const fetchFromNpmRegistry = createRegClient({})
+  const fetchFromNpmRegistry = createFetchFromRegistry({})
   const res = await fetchFromNpmRegistry('https://registry.npmjs.org/is-positive')
   const metadata = await res.json()
   t.equal(metadata.name, 'is-positive')
@@ -13,7 +13,7 @@ test('fetchFromNpmRegistry', async t => {
 })
 
 test('fetchFromNpmRegistry fullMetadata', async t => {
-  const fetchFromNpmRegistry = createRegClient({ fullMetadata: true })
+  const fetchFromNpmRegistry = createFetchFromRegistry({ fullMetadata: true })
   const res = await fetchFromNpmRegistry('https://registry.npmjs.org/is-positive')
   const metadata = await res.json()
   t.equal(metadata.name, 'is-positive')
@@ -31,7 +31,7 @@ test('authorization headers are removed before redirection if the target is on a
     .get('/is-positive')
     .reply(200, { ok: true })
 
-  const fetchFromNpmRegistry = createRegClient({ fullMetadata: true })
+  const fetchFromNpmRegistry = createFetchFromRegistry({ fullMetadata: true })
   const res = await fetchFromNpmRegistry(
     'http://registry.pnpm.js.org/is-positive',
     { authHeaderValue: 'Bearer 123' }
@@ -54,7 +54,7 @@ test('authorization headers are not removed before redirection if the target is 
     .get('/is-positive-new')
     .reply(200, { ok: true })
 
-  const fetchFromNpmRegistry = createRegClient({ fullMetadata: true })
+  const fetchFromNpmRegistry = createFetchFromRegistry({ fullMetadata: true })
   const res = await fetchFromNpmRegistry(
     'http://registry.pnpm.js.org/is-positive',
     { authHeaderValue: 'Bearer 123' }
@@ -66,7 +66,7 @@ test('authorization headers are not removed before redirection if the target is 
 })
 
 test('switch to the correct agent for requests on redirect from http: to https:', async (t) => {
-  const fetchFromNpmRegistry = createRegClient({ fullMetadata: true })
+  const fetchFromNpmRegistry = createFetchFromRegistry({ fullMetadata: true })
 
   // We can test this on any endpoint that redirects from http: to https:
   const { status } = await fetchFromNpmRegistry('http://pnpm.js.org/css/main.css')
