+++ packages/fetch/tsconfig.json
@@ -11,6 +11,9 @@
   "references": [
     {
       "path": "../core-loggers"
+    },
+    {
+      "path": "../npm-registry-agent"
     }
   ]
 }
