+++ packages/npm-resolver/package.json
@@ -35,12 +35,12 @@
   },
   "dependencies": {
     "@pnpm/error": "workspace:1.2.0",
+    "@pnpm/fetch": "workspace:^2.0.2",
     "@pnpm/resolve-workspace-range": "workspace:1.0.1",
     "@pnpm/resolver-base": "workspace:7.0.3",
     "@pnpm/types": "workspace:6.2.0",
     "credentials-by-uri": "2.0.0",
     "encode-registry": "2.0.2",
-    "fetch-from-npm-registry": "workspace:4.1.2",
     "load-json-file": "6.2.0",
     "mem": "^6.1.0",
     "mz": "^2.7.0",
