+++ packages/npm-resolver/tsconfig.json
@@ -14,6 +14,9 @@
       "path": "../error"
     },
     {
+      "path": "../fetch"
+    },
+    {
       "path": "../resolve-workspace-range"
     },
     {
@@ -21,9 +24,6 @@
     },
     {
       "path": "../types"
-    },
-    {
-      "path": "../fetch-from-npm-registry"
     }
   ]
 }
