+++ packages/outdated/package.json
@@ -38,6 +38,7 @@
     "@pnpm/constants": "workspace:4.0.0",
     "@pnpm/default-resolver": "workspace:9.0.3",
     "@pnpm/error": "workspace:1.2.0",
+    "@pnpm/fetch": "workspace:^2.0.2",
     "@pnpm/lockfile-file": "workspace:3.0.11",
     "@pnpm/lockfile-utils": "workspace:2.0.15",
     "@pnpm/manifest-utils": "workspace:1.0.3",
@@ -47,7 +48,6 @@
     "@pnpm/store-path": "^4.0.0",
     "@pnpm/types": "workspace:6.2.0",
     "dependency-path": "workspace:5.0.2",
-    "fetch-from-npm-registry": "workspace:^4.1.2",
     "lru-cache": "^5.1.1",
     "ramda": "^0.27.0",
     "semver": "^7.3.2"
