+++ packages/outdated/src/createManifestGetter.ts
@@ -1,7 +1,7 @@
 import createResolver, { ResolveFunction, ResolverFactoryOptions } from '@pnpm/default-resolver'
+import { createFetchFromRegistry } from '@pnpm/fetch'
 import pickRegistryForPackage from '@pnpm/pick-registry-for-package'
 import { DependencyManifest, Registries } from '@pnpm/types'
-import fetchFromNpmRegistry from 'fetch-from-npm-registry'
 import LRU = require('lru-cache')
 
 type GetManifestOpts = {
@@ -15,7 +15,7 @@ export type ManifestGetterOptions = Omit<ResolverFactoryOptions, 'metaCache'> & 
 export function createManifestGetter (
   opts: ManifestGetterOptions
 ): (packageName: string, pref: string) => Promise<DependencyManifest | null> {
-  const fetch = fetchFromNpmRegistry(opts)
+  const fetch = createFetchFromRegistry(opts)
   const resolve = createResolver(fetch, Object.assign(opts, {
     metaCache: new LRU({
       max: 10000,
