+++ packages/outdated/tsconfig.json
@@ -19,6 +19,9 @@
       "path": "../error"
     },
     {
+      "path": "../fetch"
+    },
+    {
       "path": "../lockfile-file"
     },
     {
@@ -41,9 +44,6 @@
     },
     {
       "path": "../dependency-path"
-    },
-    {
-      "path": "../fetch-from-npm-registry"
     }
   ]
 }
