+++ packages/plugin-commands-publishing/package.json
@@ -55,6 +55,7 @@
     "@pnpm/cli-utils": "workspace:0.4.11",
     "@pnpm/config": "workspace:10.0.1",
     "@pnpm/error": "workspace:1.2.0",
+    "@pnpm/fetch": "workspace:^2.0.2",
     "@pnpm/lifecycle": "workspace:9.1.3",
     "@pnpm/npm-resolver": "workspace:8.1.2",
     "@pnpm/pick-registry-for-package": "workspace:1.0.3",
@@ -68,7 +69,6 @@
     "cp-file": "^9.0.0",
     "enquirer": "^2.3.5",
     "fast-glob": "^3.2.4",
-    "fetch-from-npm-registry": "workspace:^4.1.2",
     "lru-cache": "^5.1.1",
     "mz": "^2.7.0",
     "p-filter": "^2.1.0",
