+++ packages/plugin-commands-publishing/src/recursivePublish.ts
@@ -1,4 +1,5 @@
 import { Config } from '@pnpm/config'
+import { createFetchFromRegistry } from '@pnpm/fetch'
 import createResolver from '@pnpm/npm-resolver'
 import pickRegistryForPackage from '@pnpm/pick-registry-for-package'
 import { ResolveFunction } from '@pnpm/resolver-base'
@@ -6,7 +7,6 @@ import runNpm from '@pnpm/run-npm'
 import sortPackages from '@pnpm/sort-packages'
 import storePath from '@pnpm/store-path'
 import { Registries } from '@pnpm/types'
-import fetchFromNpmRegistry from 'fetch-from-npm-registry'
 import LRU = require('lru-cache')
 import pFilter = require('p-filter')
 import { handler as publish } from './publish'
@@ -50,7 +50,7 @@ export default async function (
 ) {
   const pkgs = Object.values(opts.selectedProjectsGraph).map((wsPkg) => wsPkg.package)
   const storeDir = await storePath(opts.workspaceDir, opts.storeDir)
-  const fetch = fetchFromNpmRegistry(opts)
+  const fetch = createFetchFromRegistry(opts)
   const resolve = createResolver(fetch, Object.assign(opts, {
     metaCache: new LRU({
       max: 10000,
