+++ packages/tarball-fetcher/package.json
@@ -36,10 +36,10 @@
   "dependencies": {
     "@pnpm/core-loggers": "workspace:^4.1.2",
     "@pnpm/error": "workspace:1.2.0",
+    "@pnpm/fetch": "workspace:^2.0.2",
     "@pnpm/fetcher-base": "workspace:8.0.2",
     "@zkochan/retry": "^0.2.0",
     "credentials-by-uri": "^2.0.0",
-    "fetch-from-npm-registry": "workspace:4.1.2",
     "graceful-fs": "^4.2.4",
     "mem": "^6.1.0",
     "mz": "^2.7.0",
